Namespace all classes in “main” sub-app, and mount it in a top-level app
This change will allow multiple sub-apps to properly live side by side without any constant or container registration namespace collisions. Adding the `Dummy::Application` also demonstrates how multiple sub-apps could be explicitly mounted within an overall top-level application.
