+++ skeletons/simple/spec/unit/user_spec.rb
@@ -1,7 +1,7 @@
-require 'entities/user'
+require 'main/entities/user'
 
-RSpec.describe Entities::User do
-  subject(:user) { Entities::User.new(1, 'Jane') }
+RSpec.describe Main::Entities::User do
+  subject(:user) { Main::Entities::User.new(1, 'Jane') }
 
   describe '#id' do
     it 'works' do
