--- spec/dummy/apps/main/lib/entities/user.rb
@@ -1,3 +1,0 @@
-module Entities
-  User = Struct.new(:id, :name)
-end
