+++ spec/dummy/apps/main/lib/main/entities/user.rb
@@ -1,0 +1,5 @@
+module Main
+  module Entities
+    User = Struct.new(:id, :name)
+  end
+end
