+++ spec/dummy/apps/main/lib/main/transactions/register_user.rb
@@ -1,0 +1,19 @@
+require 'main/import'
+require 'main/entities/user'
+require 'transaction'
+
+module Main
+  module Transactions
+    class RegisterUser < Transaction
+      include Main::Import(:db)
+
+      def call(params)
+        if params['name']
+          success(db[:users] << Main::Entities::User.new(*params.values_at('id', 'name')))
+        else
+          failure(:validation, 'name is missing')
+        end
+      end
+    end
+  end
+end
