+++ spec/dummy/apps/main/lib/main/views/users/index.rb
@@ -1,0 +1,17 @@
+require 'main/view'
+
+module Main
+  module Views
+    module Users
+      class Index < Main::View
+        configure do |config|
+          config.template = 'users/index'
+        end
+
+        def locals(options)
+          { users: [{ name: 'Jane' }, { name: 'Joe' }] }
+        end
+      end
+    end
+  end
+end
