--- spec/dummy/apps/main/lib/views/users/index.rb
@@ -1,15 +1,0 @@
-require 'main/view'
-
-module Views
-  module Users
-    class Index < Main::View
-      configure do |config|
-        config.template = 'users/index'
-      end
-
-      def locals(options)
-        { users: [{ name: 'Jane' }, { name: 'Joe' }] }
-      end
-    end
-  end
-end
