+++ spec/dummy/config.ru
@@ -1,3 +1,3 @@
 require_relative 'core/boot'
 
-run Main::Application.freeze.app
+run Dummy::Application.freeze.app
