+++ spec/dummy/core/boot.rb
@@ -13,3 +13,5 @@ end
 
 app_paths = Pathname(__FILE__).dirname.join('../apps').realpath.join('*')
 Dir[app_paths].each { |f| require "#{f}/core/boot" }
+
+require_relative "dummy/application"
