+++ spec/dummy/core/dummy/application.rb
@@ -1,0 +1,7 @@
+module Dummy
+  class Application < Roda
+    route do |r|
+      r.run Main::Application.freeze.app
+    end
+  end
+end
